Create css with only calendar widget styles
The generated application.css contained generic element styles used on
the demo site. This interfered with a site's styles when trying to use
this widget.

This PR creates a second compiled file "calendar.css" that only
contains the date picker styles and updates the bower.json to point to
this file.