Added compatibility with Moment Timezone.
This allows the use of this component with pages that use Moment Timezone and have a default timezone set. Without this fix, the component interprets the dates from the input fields as UTC, causing Calendar.checkDate to return erroneous results, which are then put back into the input fields as timezoneless dates. This causes some dates to be off by days.