implement proof-of-concept for preserving selected preset labels
When selecting a preset from the preset dropdown, instead of being immediately translated into its date range representation, the preset label is now displayed inside the component.

The idea is that the component consumer can now work with this preset differently, the primary use case being that e.g. "Last week" can be treated symbolic and updated when the page is reloaded on a Sunday/Monday and the date range representation of the preset has changed.

The moment the user clicks on the displayed preset, it *is* turned into its date range representation and the user can set a custom date range, taking the preset's date range as the starting point.